resolve some conversations
