+++ blotztask-ui/src/model/add-task-item-dto.ts
@@ -1,0 +1,6 @@
+export interface AddTaskTtemDTO {
+  title: string;
+  description: string;
+  dueDate: string;
+  labelId: number;
+}
